Prefer go-jsonnet to jsonnet
Google has a newer implementation of jsonnet, written (mostly) in Go:

https://github.com/google/go-jsonnet

It's compatible with the original jsonnet, has a few additional
features, and is "orders of magnitude faster [in some cases]."

asdf will install this as "jsonnet", not "go-jsonnet", so this change
should transparently work with our existing jsonnet Makefile module.
However, the default go-jsonnet asdf plugin uses Go to compile the
binary, rather than download the pre-built binary. A comment has been
added to the tool-versions file to work around this if you don't want to
install Go.