Fix yq tool version
There is a bug in the asdf plugin for yq where recent versions of yq
won't install if the version number is prefixed with a `v`. (It has to
do with the yq repo changing how they format release download URLs in
GitHub.)

The error only occurs the first time you install that version yq.
Thereafter a version prefixed with `v` will work because asdf detects
the version as already being installed.