Run jsonnet/install before tanka/generate
It's all too easy to pull some changes to a `jsonnetfile.lock.json` file
and not realize to need to run `make jsonnet/install` to update your
local dependencies. And then you wonder why `make tanka/generate`
changed so many files.

This commit runs the `jsonnet/install` target as a dependency of the
`tanka/generate` target. If there are no dependencies to update, the
build time added is pretty small.