Fix crash due to terminal colors with no terminal
We attempt to print some pretty colors in a few places of our
build-harness-extensions. However, if these Makefiles are being used in
a non-tty environment (such as renovate's postUpdateTasks) the `tput`
command will fail with an error, causing the whole make command to fail.
This commit falls back to ordinary, unformatted text when TERM is not
set or unknown.