Render manifests to a tmp dir first
The way our manifest generation script currently works is:

1. Delete the old manifests
2. Render the new ones

If there is an error rendering one app, the script will exit and the
manifests for all other apps will still have been deleted. This is
annoying to see in `git status` or renovate MR changes.

This commit changes the script to render the manifests to a temporary
directory first, and then move them into the real rendered dir.

As a bonus I think the script is now faster overall because I've added
some more parallelism.