Use Grafana datasources from core-cluster-jsonnet for local dev
This commit removes the need for the build-harness-extensions-private
repo. Instead, the grafana-operator manifests will be fetched from
core-cluster-jsonnet and patched by `yq` to perform local dashboard
development.

Additionally, if you are developing dashboards in core-cluster-jsonnet,
it won't checkout a second core-cluster-jsonnet repo to get the
grafana-operator manifests. This opens up options for testing changes to
our datasources locally.